Fixing Profile and refactoring Users and Projects
